Clean up answerfile settings
A few items related to the packstack answerfile can be cleaned up
now that some changes in packstack have merged upstream.

The default token format is now UUID, so we don't need to explicitly
set it (LP#1382160).

We need to use sed to replace the existing Keystone service setting
in the answerfile now that the setting exists there by default.