Allow nova to validate tokens using the v3 API
Nova is validating tokens using the v2 API by default.  This breaks
multi-domain setups, as the v2 API only validates tokens for the
default domain.

This patch adds a workaround that configures nova to use the v3 API
for token validation.  A bug has been opened to change nova's default
configuration.  When this bug is fixed, the workaround can be removed.