Fix race condition with using newly created image for Heat
The multi-domain setup creates a new Cirros image that downloads
the image from an external URL.  The automation then attempts to
use this image to create a Heat stack.  A race condition exists
where the image may not be downloaded and active yet, which can
cause the Heat stack create to fail.

This patch adds logic to check if the image is active before
proceeding.  We timeout after one minute to prevent hanging
forever.