Update federation setup to use recent Ipsilon packages
This updates the federation setup to use a more recent Ipsilon
release.  The packaging layout changes, so there are new subpackages
that we need to install.  We also configure Ipsilon to map kerberos
principals to local usernames, which allows the 'admin' user to use
either form-based or Kerberos authentication for Ipsilon administration
tasks.

In addition, we no longer need to install OSC from git.  We can
simply use the rpms from RDO.  This removes the installation of
OSC from git and also cleans up some of the build dependencies
that we no longer need.