Check if federation extension is already in paste pipeline
In Kilo, the default Keystone paste.ini akready has the federation
extension in it.  We should check if it's already set before attemting
to add it.

Also, we were previously setting the entire api_v3 pipeline value
just to add the federation extenstion.  Since the defaults change
over time, this can cause of to modify other values that were not
intended to be changed.  We now read the original pipeline and only
add the federation extension, preserving any other values.