Configure webSSO for federation setup
This patch allows Keystone's WebSSO functionality to be optionally
configured.  This currently only works when Delorean packages are
being used since it relies on functionality that is going into
the Kilo OpenStack release.

At the end of the setup scripts, you have to add a user to IPA,
then do a kinit at that user to reset the expired password.  At
that point, a browser can be used to access the following URL:

  http://rdo.rdodom.test:5000/v3/auth/OS-FEDERATION/websso/saml2

You will need to add the following query param to the above URL to
simulate a trusted dashboard:

  ?origin=http%3A//rdo.rdodom.test

This will trigger the WebSSO flow, which looks like this:

  keystone -> IdP -> keystone (token issued) -> horizon

Horizon is not yet configure for WebSSO in this setup, but you can
use something like Firefox's "Web Console" developer tool to see
that keystone issues a redirect to Horizon and does a POST of a form
that contains a keystone token.